requirements.txt
streamlit>=1.32
pandas
psutil
dnspython
python-whois
